# it-helpdesk-agent/openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IT Ticket Manager API",
    "version": "1.0.0",
    "description": "API for managing IT support tickets. This tool allows agents to create, read, update, and delete IT support tickets including incidents, access requests, and general requests."
  },
  "servers": [
    {
      "url": "https://nervily-chordamesodermic-renda.ngrok-free.dev",
      "description": "Development server (replace with your ngrok URL)"
    }
  ],
  "paths": {
...
    "/api/tickets/{id}": {
      "get": {
        "summary": "Get a specific ticket",
        "description": "Retrieve details of a specific ticket by its ID. Use this to check the status or details of a previously created ticket.",
        "operationId": "getTicketById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique ticket ID (e.g., TKT-ABC123)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a ticket",
        "description": "Update an existing ticket's details. Use this to change the status (e.g., from 'new' to 'in_progress'), update priority, or add information to the ticket.",
        "operationId": "updateTicket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique ticket ID to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketRequest"
              },
              "example": {
                "status": "in_progress",
                "priority": "high"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a ticket",
        "description": "Permanently delete a ticket from the system. Use with caution - this action cannot be undone.",
        "operationId": "deleteTicket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique ticket ID to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Ticket": {
        "type": "object",
        "required": [
          "id",
          "ticket_type",
          "title",
          "description",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique ticket identifier",
            "example": "TKT-ABC123"
          },
          "ticket_type": {
            "type": "string",
            "enum": ["incident", "access_request", "request"],
            "description": "Type of ticket: incident (something broken), access_request (permission needed), or request (general help)"
          },
          "title": {
            "type": "string",
            "description": "Short summary of the issue or request",
            "example": "Cannot access email on laptop"
          },
          "description": {
            "type": "string",
            "description": "Detailed description with bullet points",
            "example": "- User reports Outlook keeps crashing\n- Issue started this morning"
          },
          "user_name": {
            "type": "string",
            "description": "Name of the user who reported the issue",
            "example": "John Smith"
          },
          "app_or_system": {
            "type": "string",
            "description": "The affected application or system",
            "example": "Outlook"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "description": "Ticket priority level"
          },
          "collected_details": {
            "type": "object",
            "description": "Additional structured data about the issue",
            "additionalProperties": true
          },
          "suggested_team": {
            "type": "string",
            "enum": [
              "Network",
              "Endpoint Support",
              "Application Support",
              "IAM",
              "Security",
              "DevOps"
            ],
            "description": "Team that should handle this ticket"
          },
          "status": {
            "type": "string",
            "enum": ["new", "open", "in_progress", "resolved", "closed"],
            "description": "Current status of the ticket"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the ticket was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the ticket was last updated"
          }
        }
      },
...
      "UpdateTicketRequest": {
        "type": "object",
        "properties": {
          "ticket_type": {
            "type": "string",
            "enum": ["incident", "access_request", "request"]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          },
          "app_or_system": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"]
          },
          "collected_details": {
            "type": "object",
            "additionalProperties": true
          },
          "suggested_team": {
            "type": "string",
            "enum": [
              "Network",
              "Endpoint Support",
              "Application Support",
              "IAM",
              "Security",
              "DevOps"
            ]
          },
          "status": {
            "type": "string",
            "enum": ["new", "open", "in_progress", "resolved", "closed"]
          }
        }
      },
      "TicketResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Ticket"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Error message explaining what went wrong"
          }
        }
      }
    }
  }
}